pyproject.toml
[project]
name = "cx-db8"
version = "2.0.0"
description = "Unsupervised, contextual, token-level extractive summarizer for competitive debate evidence"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Allen Roush", email = "allen.roush@gmail.com" },
]
keywords = ["nlp", "summarization", "debate", "embeddings", "extractive-summarization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "sentence-transformers>=3.0",
    "rich>=13.0",
    "typer>=0.12",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "syntok>=1.4",
    "python-docx>=1.0",
]

[project.optional-dependencies]
viz = [
    "matplotlib>=3.8",
    "umap-learn>=0.5",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
cx-db8 = "cx_db8.cli:app"

[project.urls]
Homepage = "https://github.com/Hellisotherpeople/CX_DB8"
Repository = "https://github.com/Hellisotherpeople/CX_DB8"
Issues = "https://github.com/Hellisotherpeople/CX_DB8/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cx_db8"]

[tool.pytest.ini_options]
testpaths = ["tests"]
